+ add all

+++ mycode/杂项/unordered_map 自写哈希.cpp
@@ -14,4 +14,4 @@ struct custom_hash {
     }
 };
 
-unordered_map<long long, int, custom_hash> safe_map;
+unordered_map<uint64_t, int, custom_hash> safe_map;
